Split DataCollector into four classes
Split up the current DataCollector into the following four classes:
1. ManifestFetcher.fetch
2. DataCollector.collect
3. DataPackages.package
4. DataUploader.upload

These can be chained together as follows:
```
Cfme::CloudServices::ManifestFetcher.fetch do |manifest|
  targets = ManageIQ::Providers::Vmware::InfraManager.all
  Cfme::CloudServices::DataCollector.collect(manifest, targets) do |payload|
    Cfme::CloudServices::DataPackager.package(payload) do |payload_path|
      Cfme::CloudServices::DataUploader.upload(payload_path)
      payload_path.unlink
    end
  end
end
```
